Add query strings to proxy request
All the major sites that are using notion (super.so) with custom domains are currently down because of a change notion made to images they host. They now require query string identifiers at the end of the image request. The current code for your notionUrl doesn't account for that so I changed it.